# apps/api/openapi/mimorii.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/me": {
      "get": {
        "operationId": "current",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/teams/{teamId}/resources": {
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Resources"
        ]
      },
      "post": {
        "operationId": "create",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResourceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Resources"
        ]
      }
    },
    "/api/teams/{teamId}/resources/{id}": {
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Resources"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResourceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Resources"
        ]
      },
      "delete": {
        "operationId": "remove",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Resources"
        ]
      }
    },
    "/api/teams/{teamId}/maintenance": {
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Maintenance"
        ]
      },
      "post": {
        "operationId": "create",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMaintenanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Maintenance"
        ]
      }
    },
    "/api/teams/{teamId}/maintenance/{id}": {
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Maintenance"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMaintenanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Maintenance"
        ]
      },
      "delete": {
        "operationId": "remove",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Maintenance"
        ]
      }
    },
    "/api/teams/{teamId}/maintenance/{id}/cancel": {
      "post": {
        "operationId": "cancel",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Maintenance"
        ]
      }
    },
...
  },
  "info": {
    "title": "Mimorii API",
    "description": "Self-hosted uptime, server analytics, and relay-agent API.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "User access token"
      },
      "agent-key": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Unique agent enrollment key"
      }
    },
    "schemas": {
...
      "CreateResourceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "kind": {
            "type": "string",
            "enum": [
              "server",
              "service",
              "endpoint"
            ]
          },
          "target": {
            "type": "string",
            "maxLength": 2048
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "tags": {
            "maxItems": 12,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "agentId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          }
        },
        "required": [
          "name",
          "kind",
          "target"
        ]
      },
      "UpdateResourceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "kind": {
            "type": "string",
            "enum": [
              "server",
              "service",
              "endpoint"
            ]
          },
          "target": {
            "type": "string",
            "maxLength": 2048
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "tags": {
            "maxItems": 12,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "agentId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          }
        }
      },
      "CreateMaintenanceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "startsAt": {
            "type": "string"
          },
          "endsAt": {
            "type": "string"
          },
          "recurrence": {
            "type": "string",
            "enum": [
              "none",
              "daily",
              "weekly",
              "monthly"
            ],
            "default": "none"
          },
          "recurrenceUntil": {
            "type": "object",
            "nullable": true
          },
          "suppressNotifications": {
            "type": "boolean",
            "default": true
          },
          "resourceIds": {
            "minItems": 1,
            "maxItems": 200,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "startsAt",
          "endsAt",
          "resourceIds"
        ]
      },
      "UpdateMaintenanceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "startsAt": {
            "type": "string"
          },
          "endsAt": {
            "type": "string"
          },
          "recurrence": {
            "type": "string",
            "enum": [
              "none",
              "daily",
              "weekly",
              "monthly"
            ],
            "default": "none"
          },
          "recurrenceUntil": {
            "type": "object",
            "nullable": true
          },
          "suppressNotifications": {
            "type": "boolean",
            "default": true
          },
          "resourceIds": {
            "minItems": 1,
            "maxItems": 200,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}